requirements.txt
setuptools~=65.5.0
pandas~=2.2.0
dagster~=1.6.5
SQLAlchemy~=2.0.25
boto3~=1.34.45
python-dotenv~=1.0.1
numpy~=1.26.3